requirements.txt
tensorflow==2.15.0
flask
gunicorn
numpy
pillow
scikit-learn
matplotlib
seaborn
